Refactor the code for making the functions pure
