+++ src/display.js
@@ -1,4 +1,3 @@
-import tasksList from './data.js';
 import updateStatus from './status.js';
 import taskManager from './taskManager.js';
 
@@ -26,8 +25,9 @@ class DisplayElements {
     const removeBtn = listElement.querySelector('.remove-icon');
 
     removeBtn.addEventListener('click', () => {
-      taskManager.removeTask(index);
-      this.displayElements();
+      const tasksList = JSON.parse(localStorage.getItem('todotask')) || [];
+      taskManager.removeTask(index, tasksList);
+      this.displayElements(tasksList);
     });
 
     // * Hundle task list onfocus event
@@ -40,25 +40,26 @@ class DisplayElements {
     // * Hundle task list focus lost event
 
     taskInputElement.addEventListener('blur', () => {
+      const tasksList = JSON.parse(localStorage.getItem('todotask')) || [];
       if (taskInputElement.value === '') {
-        this.displayElements();
+        this.displayElements(tasksList);
         return;
       }
       listElement.classList.remove('transform-input');
-      taskManager.updateTask(taskInputElement.value, index);
-      this.displayElements();
+      taskManager.updateTask(taskInputElement.value, index, tasksList);
     });
 
     const checkboxElement = listElement.querySelector('.checkboxInput');
 
     checkboxElement.addEventListener('change', () => {
-      updateStatus(index, checkboxElement.checked);
+      const tasksList = JSON.parse(localStorage.getItem('todotask')) || [];
+      updateStatus(index, checkboxElement.checked, tasksList);
     });
 
     return listElement;
   }
 
-  displayElements = () => {
+  displayElements = (tasksList) => {
     const elementContainer = document.querySelector('.list-box');
     elementContainer.innerHTML = '';
     tasksList.forEach((task, index) => {
