+++ src/index.html
@@ -16,7 +16,7 @@
     <header class="section-header">
       <div class="title">To Do</div>
       <div class="refresh-box">
-        <a href="#"><ion-icon class="icon" name="refresh"></ion-icon></a>
+        <ion-icon class="icon" name="refresh"></ion-icon>
       </div>
     </header>
     <!-- main section -->
