+++ src/status.js
@@ -1,13 +1,12 @@
-import tasksList from './data.js';
 import updateLocalStorage from './localStorage.js';
 
-const updateStatus = (index, status) => {
+const updateStatus = (index, status, tasksList) => {
   if (status) {
     tasksList[index].completed = true;
   } else {
     tasksList[index].completed = false;
   }
-  updateLocalStorage();
+  updateLocalStorage(tasksList);
 };
 
 export default updateStatus;
